Pawel Work in Progress including AllocationDict, CommonLabels and conflicliting BoxNumeric moved to allocations.py
