+++ tests/test_allocations.py
@@ -14,306 +14,149 @@
 
 import pytest
 
-from owca.allocators import _calculate_task_allocations_changeset, \
-    _calculate_tasks_allocations_changeset, AllocationType, \
-    _convert_tasks_allocations_to_metrics, BoxedNumeric, BoxedAllocationFactory
-from owca.metrics import Metric, MetricType
-from owca.resctrl import RDTAllocation, _parse_schemata_file_row, _count_enabled_bits, \
-    check_cbm_bits
-
 
 @pytest.mark.parametrize(
-    'current_task_allocations,new_task_allocations,'
-    'expected_target_task_allocations,expected_task_allocations_changeset', (
-        ({}, {},
-         {}, {}),
-        ({'a': 0.2}, {},
-         {'a': 0.2}, {}),
-        ({'a': 0.2}, {'a': 0.2},
-         {'a': 0.2}, {}),
-        ({'b': 2}, {'b': 3},
-         {'b': 3}, {'b': 3}),
-        ({'a': 0.2, 'b': 0.4}, {'a': 0.2, 'b': 0.5},
-         {'a': 0.2, 'b': 0.5}, {'b': 0.5}),
-        ({}, {'a': 0.2, 'b': 0.5},
-         {'a': 0.2, 'b': 0.5}, {'a': 0.2, 'b': 0.5}),
-        # RDTAllocations
-        ({}, {"rdt": RDTAllocation(name='', l3='ff')},
-         {"rdt": RDTAllocation(name='', l3='ff')}, {"rdt": RDTAllocation(name='', l3='ff')}),
-        ({"rdt": RDTAllocation(name='', l3='ff')}, {},
-         {"rdt": RDTAllocation(name='', l3='ff')}, {}),
-        ({"rdt": RDTAllocation(name='', l3='ff')}, {"rdt": RDTAllocation(name='x', l3='ff')},
-         {"rdt": RDTAllocation(name='x', l3='ff')}, {"rdt": RDTAllocation(name='x', l3='ff')}),
-        ({"rdt": RDTAllocation(name='x', l3='ff')}, {"rdt": RDTAllocation(name='x', l3='dd')},
-         {"rdt": RDTAllocation(name='x', l3='dd')}, {"rdt": RDTAllocation(name='x', l3='dd')}),
-        ({"rdt": RDTAllocation(name='x', l3='dd', mb='ff')},
-         {"rdt": RDTAllocation(name='x', mb='ff')},
-         {"rdt": RDTAllocation(name='x', l3='dd', mb='ff')}, {}),
-    ))
-def test_calculate_task_allocations(
-        current_task_allocations, new_task_allocations,
-        expected_target_task_allocations, expected_task_allocations_changeset):
-    target_task_allocations, task_allocations_changeset = _calculate_task_allocations_changeset(
-        current_task_allocations, new_task_allocations
-    )
-    assert target_task_allocations == expected_target_task_allocations
-    assert task_allocations_changeset == expected_task_allocations_changeset
-
-
-@pytest.mark.parametrize(
-    'current_tasks_allocations,new_tasks_allocations,'
-    'expected_target_tasks_allocations,expected_tasks_allocations_changeset', (
-        ({}, {},
-         {}, {}),
-        (dict(t1={'a': 2}), {},
-         dict(t1={'a': BoxedNumeric(2)}), {}),
-        (dict(t1={'a': 2}), dict(t1={'a': 2.01}),  # small enough to ignore
-         dict(t1={'a': BoxedNumeric(2)}), {}),
-        (dict(t1={'a': 2}), dict(t1={'a': 2.1}),  # big enough to notice
-         dict(t1={'a': BoxedNumeric(2.1)}), dict(t1={'a': BoxedNumeric(2.1)})),
-        (dict(t1={'a': 2}), dict(t1={'a': 2}),
-         dict(t1={'a': 2}), {}),
-        (dict(t1={'a': 1}), dict(t1={'b': 2}, t2={'b': 3}),
-         dict(t1={'a': 1, 'b': 2}, t2={'b': 3}), dict(t1={'b': 2}, t2={'b': 3})),
-    ))
-def test_calculate_tasks_allocations_changeset(
-        current_tasks_allocations, new_tasks_allocations,
-        expected_target_tasks_allocations, expected_tasks_allocations_changeset
-):
-    target_tasks_allocations, tasks_allocations_changeset = _calculate_tasks_allocations_changeset(
-        current_tasks_allocations, new_tasks_allocations
-    )
-    assert target_tasks_allocations == expected_target_tasks_allocations
-    assert tasks_allocations_changeset == expected_tasks_allocations_changeset
-
-
-@pytest.mark.parametrize('hexstr,expected_bits_count', (
-        ('', 0),
-        ('1', 1),
-        ('2', 1),
-        ('3', 2),
-        ('f', 4),
-        ('f0', 4),
-        ('0f0', 4),
-        ('ff0', 8),
-        ('f1f', 9),
-        ('fffff', 20),
-))
-def test_count_enabled_bits(hexstr, expected_bits_count):
-    got_bits_count = _count_enabled_bits(hexstr)
-    assert got_bits_count == expected_bits_count
-
-
-@pytest.mark.parametrize('line,expected_domains', (
-        ('', {}),
-        ('x=2', {'x': '2'}),
-        ('x=2;y=3', {'x': '2', 'y': '3'}),
-        ('foo=bar', {'foo': 'bar'}),
-        ('mb:1=20;2=50', {'1': '20', '2': '50'}),
-        ('mb:xxx=20mbs;2=50b', {'xxx': '20mbs', '2': '50b'}),
-        ('l3:0=20;1=30', {'1': '30', '0': '20'}),
-))
-def test_parse_schemata_file_row(line, expected_domains):
-    got_domains = _parse_schemata_file_row(line)
-    assert got_domains == expected_domains
-
-
-@pytest.mark.parametrize('invalid_line,expected_message', (
-        ('x=', 'value cannot be empty'),
-        ('x=2;x=3', 'Conflicting domain id found!'),
-        ('=2', 'domain_id cannot be empty!'),
-        ('2', 'Value separator is missing "="!'),
-        (';', 'domain cannot be empty'),
-        ('xxx', 'Value separator is missing "="!'),
-))
-def test_parse_invalid_schemata_file_domains(invalid_line, expected_message):
-    with pytest.raises(ValueError, match=expected_message):
-        _parse_schemata_file_row(invalid_line)
-
-
-def rdt_metric_func(type, value, **labels):
-    """Helper to create RDT like metric"""
-    return Metric(
-        name='allocation',
-        type=MetricType.GAUGE,
-        value=value,
-        labels=dict(allocation_type=type, **(labels or dict()))
-    )
-
-
-@pytest.mark.parametrize('rdt_allocation, expected_metrics', (
-    (RDTAllocation(), []),
-    (RDTAllocation(mb='mb:0=20'), [
-        rdt_metric_func('rdt_mb', 20, group_name='', domain_id='0')
-    ]),
-    (RDTAllocation(mb='mb:0=20;1=30'), [
-        rdt_metric_func('rdt_mb', 20, group_name='', domain_id='0'),
-        rdt_metric_func('rdt_mb', 30, group_name='', domain_id='1'),
-    ]),
-    (RDTAllocation(l3='l3:0=ff'), [
-        rdt_metric_func('rdt_l3_cache_ways', 8, group_name='', domain_id='0'),
-        rdt_metric_func('rdt_l3_mask', 255, group_name='', domain_id='0'),
-    ]),
-    (RDTAllocation(name='be', l3='l3:0=ff', mb='mb:0=20;1=30'), [
-        rdt_metric_func('rdt_l3_cache_ways', 8, group_name='be', domain_id='0'),
-        rdt_metric_func('rdt_l3_mask', 255, group_name='be', domain_id='0'),
-        rdt_metric_func('rdt_mb', 20, group_name='be', domain_id='0'),
-        rdt_metric_func('rdt_mb', 30, group_name='be', domain_id='1'),
-    ]),
-))
-def test_rdt_allocation_generate_metrics(rdt_allocation: RDTAllocation, expected_metrics):
-    got_metrics = rdt_allocation.generate_metrics()
-    assert got_metrics == expected_metrics
-
-
-@pytest.mark.parametrize('tasks_allocations,expected_metrics', (
-    ({}, []),
-    ({'some_task': {AllocationType.SHARES: 0.5}}, [
-        Metric(name='allocation', value=0.5,
-               type=MetricType.GAUGE,
-               labels={'allocation_type': 'cpu_shares', 'task_id': 'some_task'})
-    ]),
-    ({'some_task': {AllocationType.RDT: RDTAllocation(mb='mb:0=20')}}, [
-        rdt_metric_func('rdt_mb', 20, group_name='', domain_id='0', task_id='some_task')
-    ]),
-    ({'some_task': {AllocationType.SHARES: 0.5,
-                    AllocationType.RDT: RDTAllocation(mb='mb:0=20')}}, [
-        Metric(
-            name='allocation', value=0.5,
-            type=MetricType.GAUGE,
-            labels={'allocation_type': AllocationType.SHARES, 'task_id': 'some_task'}
-        ),
-        rdt_metric_func('rdt_mb', 20, group_name='', domain_id='0', task_id='some_task')
-    ]),
-    ({'some_task_a': {
-        AllocationType.SHARES: 0.5, AllocationType.RDT: RDTAllocation(mb='mb:0=30')
-    },
-         'some_task_b': {
-             AllocationType.QUOTA: 0.6,
-             AllocationType.RDT: RDTAllocation(name='b', l3='l3:0=f;1=f1'),
-         }}, [
-         Metric(
-             name='allocation', value=0.5,
-             type=MetricType.GAUGE,
-             labels={'allocation_type': AllocationType.SHARES, 'task_id': 'some_task_a'}
-         ),
-         rdt_metric_func('rdt_mb', 30, group_name='', domain_id='0', task_id='some_task_a'),
-         Metric(
-             name='allocation', value=0.6,
-             type=MetricType.GAUGE,
-             labels={'allocation_type': AllocationType.QUOTA, 'task_id': 'some_task_b'}
-         ),
-         rdt_metric_func('rdt_l3_cache_ways', 4, group_name='b',
-                         domain_id='0', task_id='some_task_b'),
-         rdt_metric_func('rdt_l3_mask', 15, group_name='b',
-                         domain_id='0', task_id='some_task_b'),
-         rdt_metric_func('rdt_l3_cache_ways', 5, group_name='b',
-                         domain_id='1', task_id='some_task_b'),
-         rdt_metric_func('rdt_l3_mask', 241, group_name='b',
-                         domain_id='1', task_id='some_task_b'),
-     ]),
-))
-def test_convert_task_allocations_to_metrics(tasks_allocations, expected_metrics):
-    metrics_got = _convert_tasks_allocations_to_metrics(tasks_allocations)
-    assert metrics_got == expected_metrics
-
-
-@pytest.mark.parametrize(
-    'current_rdt_alloaction, new_rdt_allocation,'
-    'expected_target_rdt_allocation,expected_rdt_allocation_changeset', (
-        (None, RDTAllocation(),
-         RDTAllocation(), RDTAllocation()),
-        (RDTAllocation(name=''), RDTAllocation(),  # empty group overrides existing
-         RDTAllocation(), RDTAllocation()),
-        (RDTAllocation(), RDTAllocation(l3='x'),
-         RDTAllocation(l3='x'), RDTAllocation(l3='x')),
-        (RDTAllocation(l3='x'), RDTAllocation(mb='y'),
-         RDTAllocation(l3='x', mb='y'), RDTAllocation(mb='y')),
-        (RDTAllocation(l3='x'), RDTAllocation(l3='x', mb='y'),
-         RDTAllocation(l3='x', mb='y'), RDTAllocation(mb='y')),
-        (RDTAllocation(l3='x', mb='y'), RDTAllocation(name='new', l3='x', mb='y'),
-         RDTAllocation(name='new', l3='x', mb='y'), RDTAllocation(name='new', l3='x', mb='y')),
-        (RDTAllocation(l3='x'), RDTAllocation(name='', l3='x'),
-         RDTAllocation(name='', l3='x'), RDTAllocation(name='', l3='x'))
-    )
+    'current_allocations, new_allocations, '
+    'expected_target_allocations, expected_allocations_changeset', [
+        ({}, {}, {}, None),
+        (dict(x=2), {}, dict(x=2), None),
+    ]
 )
-def test_merge_rdt_allocations1(
-        current_rdt_alloaction, new_rdt_allocation,
-        expected_target_rdt_allocation, expected_rdt_allocation_changeset):
-    got_target_rdt_allocation, got_rdt_alloction_changeset = \
-        new_rdt_allocation.merge_with_current(current_rdt_alloaction)
+def test_allocations_dict_merging(current_allocations, new_allocations, 
+                                  expected_target_allocations, expected_allocations_changeset
+    ):
+    got_target_allocations, got_allocations_changeset = \
+        AllocationsDict(new_allocations).merge_with_current(AllocationsDict(current_allocations))
 
-    assert got_target_rdt_allocation == expected_target_rdt_allocation
-    assert got_rdt_alloction_changeset == expected_rdt_allocation_changeset
+    expected_allocations_changeset_dict = AllocationsDict(expected_allocations_changeset) if \
+        expected_allocations_changeset is not None else None
+     
+    assert got_target_allocations == AllocationsDict(expected_target_allocations)
+    assert got_allocations_changeset == expected_allocations_changeset_dict
 
+# single task
 
-def test_check_cbm_bits_valid():
-    check_cbm_bits('ff00', 'ffff', '1')
-
-
-@pytest.mark.parametrize(
-    'mask, cbm_mask, min_cbm_bits, expected_error_message', (
-            ('f0f', 'ffff', '1', 'without a gap'),
-            ('0', 'ffff', '1', 'minimum'),
-            ('ffffff', 'ffff', 'bigger', ''),
-    )
-)
-def test_check_cbm_bits_gap(mask: str, cbm_mask: str, min_cbm_bits: str,
-                            expected_error_message: str):
-    with pytest.raises(ValueError, match=expected_error_message):
-        check_cbm_bits(mask, cbm_mask, min_cbm_bits)
-
-
-@pytest.mark.parametrize(
-    'value, min_value, max_value, float_value_change_sensitivity, is_valid', (
-            (1, 2, 3, 0.00001, ['1 does not belong to range <2;3>']),
-            (1.1, 2, 3, 0.00001, ['1.1 does not belong to range <2;3>']),
-            (2.5, 2, 3, 0.00001, []),
-            (3, 2.5, 3.0, 0.00001, []),
-            (2.0, 2, 3.0, 0.00001, []),
-            (2.0, None, 3.0, 0.00001, []),
-            (2.0, 1, None, 0.00001, []),
-    )
-)
-def test_boxed_numeric_validation(value, min_value, max_value, float_value_change_sensitivity,
-                                  is_valid):
-    boxed_value = BoxedNumeric(value, min_value, max_value, float_value_change_sensitivity)
-    assert boxed_value.validate() == is_valid
-
-
-@pytest.mark.parametrize(
-    'current_value, new_value, expected_value', (
-            (BoxedNumeric(10), BoxedNumeric(10.1), (BoxedNumeric(10), None)),
-            (BoxedNumeric(10), BoxedNumeric(10.99), (BoxedNumeric(10.99), BoxedNumeric(10.99))),
-    )
-)
-def test_boxed_numeric_merging(current_value, new_value, expected_value):
-    value, changeset = current_value.merge_with_current(new_value)
-    assert value == expected_value[0]
-    assert changeset == expected_value[1]
+# @pytest.mark.parametrize(
+#     'current_task_allocations,new_task_allocations,'
+#     'expected_target_task_allocations,expected_task_allocations_changeset', (
+#         ({}, {},
+#          {}, {}),
+#         ({'a': 0.2}, {},
+#          {'a': 0.2}, {}),
+#         ({'a': 0.2}, {'a': 0.2},
+#          {'a': 0.2}, {}),
+#         ({'b': 2}, {'b': 3},
+#          {'b': 3}, {'b': 3}),
+#         ({'a': 0.2, 'b': 0.4}, {'a': 0.2, 'b': 0.5},
+#          {'a': 0.2, 'b': 0.5}, {'b': 0.5}),
+#         ({}, {'a': 0.2, 'b': 0.5},
+#          {'a': 0.2, 'b': 0.5}, {'a': 0.2, 'b': 0.5}),
+#         # RDTAllocations
+#         ({}, {"rdt": RDTAllocation(name='', l3='ff')},
+#          {"rdt": RDTAllocation(name='', l3='ff')}, {"rdt": RDTAllocation(name='', l3='ff')}),
+#         ({"rdt": RDTAllocation(name='', l3='ff')}, {},
+#          {"rdt": RDTAllocation(name='', l3='ff')}, {}),
+#         ({"rdt": RDTAllocation(name='', l3='ff')}, {"rdt": RDTAllocation(name='x', l3='ff')},
+#          {"rdt": RDTAllocation(name='x', l3='ff')}, {"rdt": RDTAllocation(name='x', l3='ff')}),
+#         ({"rdt": RDTAllocation(name='x', l3='ff')}, {"rdt": RDTAllocation(name='x', l3='dd')},
+#          {"rdt": RDTAllocation(name='x', l3='dd')}, {"rdt": RDTAllocation(name='x', l3='dd')}),
+#         ({"rdt": RDTAllocation(name='x', l3='dd', mb='ff')}, {"rdt": RDTAllocation(name='x', mb='ff')},
+#          {"rdt": RDTAllocation(name='x', l3='dd', mb='ff')}, {}),
+#     ))
+# def test_calculate_task_allocations(
+#         current_task_allocations, new_task_allocations,
+#         expected_target_task_allocations, expected_task_allocations_changeset):
+#     target_task_allocations, task_allocations_changeset = _calculate_task_allocations_changeset(
+#         current_task_allocations, new_task_allocations
+#     )
+#     assert target_task_allocations == expected_target_task_allocations
+#     assert task_allocations_changeset == expected_task_allocations_changeset
 
 
-@pytest.mark.parametrize(
-    'left, right, is_equal', (
-            (BoxedNumeric(10), BoxedNumeric(10), True),
-            (BoxedNumeric(10), BoxedNumeric(11), False),
-            (BoxedNumeric(10), BoxedNumeric(10.1), True),
-            (BoxedNumeric(10), BoxedNumeric(10.11), False),
-            (BoxedNumeric(10.99), BoxedNumeric(10.99), True),
-    )
-)
-def test_boxed_numeric_equal(left, right, is_equal):
-    assert (left == right) == is_equal
+
+# # tasks
+#
+# @pytest.mark.parametrize(
+#     'current_tasks_allocations,new_tasks_allocations,'
+#     'expected_target_tasks_allocations,expected_tasks_allocations_changeset', (
+#         ({}, {},
+#          {}, {}),
+#         (dict(t1={'a': 2}), {},
+#          dict(t1={'a': 2}), {}),
+#         (dict(t1={'a': 2}), dict(t1={'a': 2.01}),  # small enough to ignore
+#          dict(t1={'a': 2}), {}),
+#         (dict(t1={'a': 2}), dict(t1={'a': 2.1}),  # big enough to notice
+#          dict(t1={'a': 2.1}), dict(t1={'a': 2.1})),
+#         (dict(t1={'a': 2}), dict(t1={'a': 2}),
+#          dict(t1={'a': 2}), {}),
+#         (dict(t1={'a': 1}), dict(t1={'b': 2}, t2={'b': 3}),
+#          dict(t1={'a': 1, 'b': 2}, t2={'b': 3}), dict(t1={'b': 2}, t2={'b': 3})),
+#     ))
+# def test_calculate_tasks_allocations_changeset(
+#         current_tasks_allocations, new_tasks_allocations,
+#         expected_target_tasks_allocations, expected_tasks_allocations_changeset
+# ):
+#     target_tasks_allocations, tasks_allocations_changeset = _calculate_tasks_allocations_changeset(
+#         current_tasks_allocations, new_tasks_allocations
+#     )
+#     assert target_tasks_allocations == expected_target_tasks_allocations
+#     assert tasks_allocations_changeset == expected_tasks_allocations_changeset
+#
+#
 
 
-def test_box_allocations_factory_known_type():
-    boxed_int = BoxedAllocationFactory.create(1, min_value=2, max_value=3)
-    assert boxed_int._value == 1
-    assert boxed_int._min_value == 2
-    assert boxed_int._max_value == 3
 
 
-def test_box_allocations_factory_unknown_type():
-    with pytest.raises(KeyError):
-        this_is_really_complex = complex('1+2j')
-        BoxedAllocationFactory.create(this_is_really_complex, min_value=2, max_value=3)
+# @pytest.mark.parametrize(
+#     'value, min_value, max_value, float_value_change_sensitivity, is_valid', (
+#             (1, 2, 3, 0.00001, ['1 does not belong to range <2;3>']),
+#             (1.1, 2, 3, 0.00001, ['1.1 does not belong to range <2;3>']),
+#             (2.5, 2, 3, 0.00001, []),
+#             (3, 2.5, 3.0, 0.00001, []),
+#             (2.0, 2, 3.0, 0.00001, []),
+#             (2.0, None, 3.0, 0.00001, []),
+#             (2.0, 1, None, 0.00001, []),
+#     )
+# )
+# def test_boxed_numeric_validation(value, min_value, max_value, float_value_change_sensitivity,
+#                                   is_valid):
+#     boxed_value = BoxedNumeric(value, min_value, max_value, float_value_change_sensitivity)
+#     assert boxed_value.validate() == is_valid
+#
+#
+# @pytest.mark.parametrize(
+#     'current_value, new_value, expected_value', (
+#             (BoxedNumeric(10), BoxedNumeric(10.1), (BoxedNumeric(10), None)),
+#             (BoxedNumeric(10), BoxedNumeric(10.99), (BoxedNumeric(10.99), BoxedNumeric(10.99))),
+#     )
+# )
+# def test_boxed_numeric_merging(current_value, new_value, expected_value):
+#     value, changeset = current_value.merge_with_current(new_value)
+#     assert value == expected_value[0]
+#     assert changeset == expected_value[1]
+#
+#
+# @pytest.mark.parametrize(
+#     'left, right, is_equal', (
+#             (BoxedNumeric(10), BoxedNumeric(10), True),
+#             (BoxedNumeric(10), BoxedNumeric(11), False),
+#             (BoxedNumeric(10), BoxedNumeric(10.1), True),
+#             (BoxedNumeric(10), BoxedNumeric(10.11), False),
+#             (BoxedNumeric(10.99), BoxedNumeric(10.99), True),
+#     )
+# )
+# def test_boxed_numeric_equal(left, right, is_equal):
+#     assert (left == right) == is_equal
+#
+#
+# def test_box_allocations_factory_known_type():
+#     boxed_int = BoxedAllocationFactory.create(1, min_value=2, max_value=3)
+#     assert boxed_int._value == 1
+#     assert boxed_int._min_value == 2
+#     assert boxed_int._max_value == 3
+#
+#
+# def test_box_allocations_factory_unknown_type():
+#     with pytest.raises(KeyError):
+#         this_is_really_complex = complex('1+2j')
+#         BoxedAllocationFactory.create(this_is_really_complex, min_value=2, max_value=3)
